Addition of parameters Q and alpha and bug fix in local search.
Addition of parameters Q and alpha.
Q is a constant that represent the amount of pheromone an ant put on the path. alpha is the parameter that shows the relative importance of the pheromone trace.

Bug fix in local search. 
The old local search could generate a path like this:

[2, 3, 4, 1, 1, 5]

Thus, beyond the path not begin and end in the same city, the old local search generate edges with distance 0. In this example is the edge {1,1}